requirements.txt
aiogram==3.13.1
gspread==6.1.2
oauth2client==4.1.3
APScheduler==3.10.4
python-dotenv==1.0.1
Pillow==10.4.0
